refactor(chat): move chat service and llm chain to new layers
